fix: /nar: handle nar-serve URLs
Old requests from nar-serve were never routed to the handler,
as harmonia complained about being unable to serialize `hash`,
as it failed to populate the NarRequest struct.

This means, the `get` function was never executed for the
"/nar/outhash-narhash.nar" case.

Additionally, the logic was a bit hard to debug, as multiple "hash"
variables meaning different things were present.

I refactored the URL handler to have two URL handlers, populating a new
`PathParams` struct with `narhash` and `outhash` fields.

I also pushed length validations into the regex that's now used in the
URL handler directly (and introduced validation for the strings to be
nixbase32)

This simplifies the logic in the URL handler a bit, and most
importantly, fixes support for URLs generated by nar-serve.

Both URLs were manually verified to now serve a NAR.