Move `NarHash` to `harmonia-protocol`, and rework hash infra
`NarHash` isn't a hashing utility concept, or even a core data structure
concept, it is a specific storage and transport layer convention, as
such it belongs in the protocol crate.

The infra reworking firstly allows for hashes to be moved out in such
ways. But it also has other benefits such as reducing boilerplate, and
perhaps better separating content from presentation.